requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14
python-dotenv==1.0.1
requests==2.32.3
vosk==0.3.45; sys_platform == 'win32'
pywhatkit==5.4
wikipedia==1.4.0
pyautogui==0.9.54
